+++ packages/remix-dev/config.ts
@@ -179,6 +179,15 @@ export interface AppConfig {
   serverPlatform?: ServerPlatform;
 
   /**
+   * Configuration of server bundles to produce. If this is defined then the
+   * top-level `serverBuildPath` value is ignored.
+   */
+  serverBundles?: {
+    serverBuildPath: string;
+    routes: string[];
+  }[];
+
+  /*
    * Whether to support Tailwind functions and directives in CSS files if
    * `tailwindcss` is installed. Defaults to `true`.
    */
@@ -357,6 +366,15 @@ export interface RemixConfig {
   serverPlatform: ServerPlatform;
 
   /**
+   * Configuration of server bundles to produce. If this is defined then the
+   * top-level `serverBuildPath` value is ignored.
+   */
+  serverBundles?: {
+    serverBuildPath: string;
+    routes: RouteManifest;
+  }[];
+
+  /**
    * Whether to support Tailwind functions and directives in CSS files if `tailwindcss` is installed.
    * Defaults to `true`.
    */
@@ -395,7 +413,7 @@ export async function readConfig(
     let appConfigModule: any;
     try {
       // shout out to next
-      // https://github.com/vercel/next.js/blob/b15a976e11bf1dc867c241a4c1734757427d609c/packages/next/server/config.ts#L748-L765
+      // https://github.com/khulnasoft/next.js/blob/b15a976e11bf1dc867c241a4c1734757427d609c/packages/next/server/config.ts#L748-L765
       if (process.env.JEST_WORKER_ID) {
         // dynamic import does not currently work inside vm which
         // jest relies on, so we fall back to require for this case
@@ -685,6 +703,24 @@ export async function resolveConfig(
     }
   }
 
+  let serverBundles = appConfig.serverBundles?.map((bundle) => {
+    // Build up the resolved `routes` including any parent routes
+    let bundleRoutes: RouteManifest = {};
+    for (let id of bundle.routes) {
+      let currentRoute: RouteManifest[string] | undefined = routes[id];
+      do {
+        bundleRoutes[currentRoute.id] = currentRoute;
+        if (currentRoute.parentId) {
+          currentRoute = routes[currentRoute.parentId];
+        } else {
+          currentRoute = undefined;
+        }
+      } while (currentRoute);
+    }
+
+    return { serverBuildPath: bundle.serverBuildPath, routes: bundleRoutes };
+  });
+
   let watchPaths: string[] = [];
   if (typeof appConfig.watchPaths === "function") {
     let directories = await appConfig.watchPaths();
@@ -799,6 +835,7 @@ export async function resolveConfig(
     serverNodeBuiltinsPolyfill,
     browserNodeBuiltinsPolyfill,
     serverPlatform,
+    serverBundles,
     mdx,
     postcss,
     tailwind,